added sauvola thresholding for future use

+++ utils/helpers.h
@@ -27,6 +27,6 @@ void convolution(
     double *kernel, int ksize, Image *image, Image *out, bool normalize);
 
 void convolution_mat(
-    double *kernel, int ksize_x, int ksize_y, Image *in, double *out);
+    double *kernel, int ksize_x, int ksize_y, Image *in, int *out);
 
 #endif
